+++ src/rx-document.ts
@@ -21,7 +21,8 @@ import {
 } from './rx-change-event';
 import {
     newRxError,
-    newRxTypeError
+    newRxTypeError,
+    RxError
 } from './rx-error';
 import {
     runPluginHooks
@@ -292,24 +293,42 @@ export const basePrototype = {
         throw pluginMissing('attachments');
     },
 
+
     /**
      * runs an atomic update over the document
      * @param function that takes the document-data and returns a new data-object
      */
     atomicUpdate(this: RxDocument, fun: Function): Promise<RxDocument> {
         this._atomicQueue = this._atomicQueue
-            .then(() => {
-                const oldData = this._dataSync$.getValue();
-                const ret = fun(clone(this._dataSync$.getValue()), this);
-                const retPromise = toPromise(ret);
-                return retPromise
-                    .then(newData => {
-                        // collection does not exist on local documents
-                        if (this.collection) {
-                            newData = this.collection.schema.fillObjectWithDefaults(newData);
+            .then(async () => {
+                let done = false;
+                // we need a hacky while loop to stay incide the chain-link of _atomicQueue
+                // while still having the option to run a retry on conflicts
+                while (!done) {
+                    const oldData = this._dataSync$.getValue();
+                    const ret = fun(clone(this._dataSync$.getValue()), this);
+                    let newData = await toPromise(ret);
+                    if (this.collection) {
+                        newData = this.collection.schema.fillObjectWithDefaults(newData);
+                    }
+                    try {
+                        await this._saveData(newData, oldData);
+                        done = true;
+                    } catch (err) {
+                        /**
+                         * conflicts cannot happen by just using RxDB in one process
+                         * There are two ways they still can appear which is
+                         * replication and multi-tab usage
+                         * Because atomicUpdate has a mutation function,
+                         * we can just re-run the mutation until there is no conflict
+                         */
+                        if ((err as RxError).parameters?.pouchDbError?.status === 409) {
+                            // pouchdb conflict error -> retrying
+                        } else {
+                            throw err;
                         }
-                        return this._saveData(newData, oldData);
-                    });
+                    }
+                }
             });
         return this._atomicQueue.then(() => this);
     },